# Mitigate wind-up problem in AP&F: prevent queue virtualStart lag