@MikeSpreitzer 2020-10-29
This is a partial mitigation of the problem (#95979).  This mitigation addresses one of the two sides of the imbalance.  This is better than nothing.

The other side of the imbalance is that queues using more than the even allocation build up a disadvantage in their `virtualStart`.  Such a queue keeps its accumulated disadvantage until that queue becomes empty.  This is a disadvantage relative to queues using less than the even allocation --- including empty queues.  Thus, a queue that transitions from empty to non-empty starts with a relative advantage compared to the queues that have a built-up disadvantage.